requirements.txt
streamlit>=1.32.0
langchain>=0.1.20
langchain-community>=0.0.33
langchain-groq>=0.0.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
unstructured[local-inference]>=0.12.6
python-dotenv>=1.0.1
pandas
lxml
pdfminer.six